Added more logging to build-script

+++ build-scripts/build.ts
@@ -3,15 +3,20 @@ import { execSync } from "child_process";
 import fs from "fs";
 import path from "path";
 
+console.log("Building JohnCG release");
+console.log();
+
 // check, wether the build script supports the os
 if (!["win32", "linux"].includes(process.platform)) {
 	throw new TypeError("Buildscript does not support this OS");
 }
 
 // load the package.json
+console.log("Reading 'package.json'");
 const package_json = JSON.parse(fs.readFileSync("package.json", "utf-8")) as { version: string; dependencies: string[] };
 
 const build_name = `JohnCG_${package_json.version}_${process.platform}`;
+console.log(`Building for target '${build_name}'`);
 
 let exec_name: string;
 switch (process.platform) {
@@ -27,26 +32,53 @@ switch (process.platform) {
 const build_dir = "dist/build";
 const release_dir = path.join("dist", build_name);
 
+console.log(`Build directory is '${build_dir}'`);
+console.log(`Release directory is '${release_dir}'`);
+console.log();
+
 // clear the build- and release-directory
+console.log("Removing build directory");
 fs.rmSync(build_dir, { recursive: true, force: true });
-fs.rmSync(release_dir, { recursive: true, force: true });
+
+if (fs.existsSync(release_dir)) {
+	console.log("Removing release directory");
+	fs.rmSync(release_dir, { recursive: true, force: true });
+}
+
+console.log("Creating building directory");
 fs.mkdirSync(build_dir, { recursive: true });
+
+console.log("Creating empty release directory");
 fs.mkdirSync(release_dir, { recursive: true });
+
+console.log();
 
 // helper-functions
 const copy_build_file = (file: string, dest?: string) => fs.copyFileSync(file, path.join(build_dir, dest ?? path.basename(file)));
 // const copy_build_dir = (dir: string, dest?: string, args?: fs.CopySyncOptions) => fs.cpSync(dir, path.join(build_dir, dest ?? path.basename(dir)), { recursive: true, ...args });
 const copy_release_file = (file: string, dest?: string) => fs.copyFileSync(file, path.join(release_dir, dest ?? path.basename(file)));
 const copy_release_dir = (dir: string, dest?: string, args?: fs.CopySyncOptions) => fs.cpSync(dir, path.join(release_dir, dest ?? path.basename(dir)), { recursive: true, ...args });
+const copy_module = (name: string) => copy_release_dir(`node_modules/${name}`, `node_modules/${name}/`);
 
 // write the version-number to config/version.ts
+const config_ts_path = "src/server/config/version.ts";
+console.log(`Writing version '${package_json.version}' to '${config_ts_path}`);
 fs.writeFileSync("src/server/config/version.ts", `// eslint-disable-next-line @typescript-eslint/naming-convention
 export const Version = "${package_json.version}";\n`);
 
+console.log();
+
 // bundle the different scripts
+console.log("Building server");
 execSync("npm run server-build");
+
+console.log("Building client");
 execSync("npm run client-build");
+
+console.log("Building templates");
 execSync("npm run templates-build");
+
+console.log("Building pandoc installer");
 execSync("npm run pandoc-build");
 
 // temporary method until there is a solution for packaging sharp
@@ -54,6 +86,7 @@ execSync("npm run pandoc-build");
 // execSync("node --experimental-sea-config sea-config.json");
 
 // get the node executable
+console.log(`Copying node executable to '${build_dir}'`);
 copy_build_file(process.execPath, exec_name);
 
 // temporary method until there is a solution for packaging sharp
@@ -62,38 +95,71 @@ copy_build_file(process.execPath, exec_name);
 // // modify the node executable
 // execSync(`npx postject dist/build/${exec_name} NODE_SEA_BLOB dist/build/sea-prep.blob --sentinel-fuse NODE_SEA_FUSE_fce680ab2cc467b6e072b8b5df1996b2`);
 
+console.log();
+console.log(`Copying files to '${release_dir}'`);
+
+console.log(`\tCopying 'config_default.yaml' to '${path.join(release_dir, "config.yaml")}'`);
 copy_release_dir("config_default.yaml", "config.yaml");
 
 // copy the file to the output
+console.log(`\tCopying node executable`);
 copy_release_file(path.join(build_dir, exec_name));
-copy_release_file(path.join(build_dir, "main.js"));
-fs.readdirSync("files").forEach((dir) => copy_release_dir(path.join("files", dir)));
-copy_release_dir(path.join(build_dir, "client"));
-const copy_module = (name: string) => {
-	copy_release_dir(`node_modules/${name}`, `node_modules/${name}/`);
 
-};
+console.log(`\tCopying server-script '${path.join(build_dir, "main.js")}'`);
+copy_release_file(path.join(build_dir, "main.js"));
+
+console.log(`\tCopying example files`);
+fs.readdirSync("files").forEach((dir) => {
+	console.log(`\t\tCopying '${path.join("files", dir)}'`);
+	copy_release_dir(path.join("files", dir));
+});
+
+console.log(`\tCopying '${path.join(build_dir, "client")}'`);
+copy_release_dir(path.join(build_dir, "client"));
+
+console.log("\tCopying CasparCG-templates");
 copy_release_dir(path.join(build_dir, "Templates"));
+
+console.log("\tCopying CasparCG-media");
 copy_release_dir("casparcg/Media");
+
+console.log("\tCopying pandoc-installer");
 copy_release_dir(path.join(build_dir, "pandoc"));
+
+console.log("\tCopying texlive-profile");
 copy_release_file("pandoc/texlive.profile", "pandoc/texlive.profile");
+
+console.log("\tCopying external node-modules");
+
+console.log("\t\tCopying '@img'");
 copy_module("@img");
+
+console.log("\t\tCopying 'canvas'");
 copy_module("canvas");
+
+console.log("\t\tCopying 'pdfjs-dist'");
 copy_module("pdfjs-dist");
+
+console.log();
 
 // temporary method until there is a solution for packaging sharp
 // create a script-file, that start node with the main.js
+console.log("Creating startup-script for the server");
 create_launch_script("main.js", build_name);
+
+console.log("Creating startup-script for the pandoc-installer");
 create_launch_script("pandoc-installer.js", "pandoc/install");
 
 // create and copy the licenses
-// void lr.cli(["--config=build-scripts/license-reporter.config.ts"]);
+console.log("Creating node-module licence-report");
 try {
 	execSync("npx license-reporter --config build-scripts/license-reporter.config.ts");
 } catch (e) { /* empty */ }
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 interface License { name: string; licenseText: string }
+
+console.log("Loading licence-report");
 const licenses_orig = JSON.parse(fs.readFileSync("build-scripts/3rdpartylicenses.json", "utf-8")) as License[];
 
 const licenses: Record<string, License> = {};
@@ -102,10 +168,14 @@ licenses_orig.forEach((pack) => {
 	licenses[pack.name] = pack;
 });
 
+console.log("Creating licence-directory");
 fs.mkdirSync("dist/build/licenses");
 
+console.log("Writing licences");
 Object.keys(package_json.dependencies).forEach((pack) => {
 	const lic = licenses[pack];
+
+	console.log(`\t'${lic.name}'`);
 
 	try {
 		fs.writeFileSync(`dist/build/licenses/${lic.name}.txt`, lic.licenseText, "utf-8");
@@ -116,12 +186,17 @@ Object.keys(package_json.dependencies).forEach((pack) => {
 	}
 });
 
+console.log("Writing JohnCG-licene")
 copy_release_file("LICENSE", "LICENSE.txt");
 
 // copy the licenses
+console.log(`Copying licences to '${release_dir}'`);
 copy_release_dir(path.join(build_dir, "licenses"));
 
+console.log();
+
 // pack the files
+console.log(`Packing release to '${release_dir}.zip'`);
 const zip_stream = fs.createWriteStream(release_dir + ".zip");
 
 const archive = archiver("zip");
